refactor: refactor the schemas structure
