+++ src/album/schemas/album.interface.ts
@@ -1,0 +1,18 @@
+export interface IAlbum {
+  id: string;
+  name: string;
+  year: number;
+  artistId: string | null;
+}
+
+export interface ICreateAlbumDto {
+  name: string;
+  year: number;
+  artistId: string | null;
+}
+
+export interface IUpdateAlbumDto {
+  name?: string;
+  year?: number;
+  artistId?: string | null;
+}
