+++ src/album/schemas/album.model.ts
