+++ src/artist/schemas/artist.interface.ts
@@ -1,0 +1,15 @@
+export interface IArtist {
+  id: string;
+  name: string;
+  grammy: boolean;
+}
+
+export interface ICreateArtistDto {
+  name: string;
+  grammy: boolean;
+}
+
+export interface IUpdateArtistDto {
+  name?: string;
+  grammy?: boolean;
+}
