+++ src/artist/schemas/artist.model.ts
