+++ src/user/schemas/user.interface.ts
@@ -1,0 +1,18 @@
+export interface IUser {
+  id: string;
+  login: string;
+  password?: string;
+  version: number;
+  createdAt: number;
+  updatedAt: number;
+}
+
+export interface ICreateUserDto {
+  login: string;
+  password: string;
+}
+
+export interface IUpdatePasswordDto {
+  oldPassword: string;
+  newPassword: string;
+}
